internal/status: move Cache methods into cache.go, inline commit funcs
