+++ internal/status/routeconditions.go
@@ -72,33 +72,6 @@ func (routeUpdate *RouteConditionsUpdate) AddCondition(cond gatewayapi_v1alpha2.
 	return newDc
 }
 
-// RouteConditionsAccessor returns a RouteConditionsUpdate that allows a client to build up a list of
-// metav1.Conditions as well as a function to commit the change back to the cache when everything
-// is done. The commit function pattern is used so that the RouteConditionsUpdate does not need
-// to know anything the cache internals.
-func (c *Cache) RouteConditionsAccessor(nsName types.NamespacedName, generation int64, resource client.Object, gateways []gatewayapi_v1alpha2.RouteParentStatus) (*RouteConditionsUpdate, func()) {
-	pu := &RouteConditionsUpdate{
-		FullName:           nsName,
-		Conditions:         make(map[gatewayapi_v1alpha2.RouteConditionType]metav1.Condition),
-		ExistingConditions: c.getRouteGatewayConditions(gateways),
-		GatewayRef:         c.gatewayRef,
-		GatewayController:  c.gatewayController,
-		Generation:         generation,
-		TransitionTime:     metav1.NewTime(clock.Now()),
-		Resource:           resource,
-	}
-
-	return pu, func() {
-		c.commitRoute(pu)
-	}
-}
-
-func (c *Cache) commitRoute(pu *RouteConditionsUpdate) {
-	if len(pu.Conditions) == 0 {
-		return
-	}
-	c.routeUpdates[pu.FullName] = pu
-}
 func (routeUpdate *RouteConditionsUpdate) Mutate(obj client.Object) client.Object {
 
 	var gatewayStatuses []gatewayapi_v1alpha2.RouteParentStatus
